+++ spotinst/ocean_aks_np_virtual_node_group_strategy/consts.go
@@ -3,7 +3,8 @@ package ocean_aks_np_virtual_node_group_strategy
 import "github.com/spotinst/terraform-provider-spotinst/spotinst/commons"
 
 const (
-	SpotPercentage     commons.FieldName = "spot_percentage"
-	FallbackToOnDemand commons.FieldName = "fallback_to_ondemand"
-	DrainingTimeout    commons.FieldName = "draining_timeout"
+	SpotPercentage           commons.FieldName = "spot_percentage"
+	FallbackToOnDemand       commons.FieldName = "fallback_to_ondemand"
+	DrainingTimeout          commons.FieldName = "draining_timeout"
+	ShouldUtilizeCommitments commons.FieldName = "should_utilize_commitments"
 )